COMPAT: blacklist numexpr=2.4.4
closes #12489
closes #12511
